refactor: RandomBallsGenerator 패키지 변경
domain -> domain.util
static 메서드만 가진 유틸성 메서드이므로 util 패키지로 분리하는 게 적절하다고 생각함
